CategoryPlot
rename com.twosigma.beaker.chart.categoryplot.plotitems package to the com.twosigma.beaker.chart.categoryplot.plotitem
