+++ plugin/jvm/src/main/java/com/twosigma/beaker/chart/categoryplot/plotitem/CategoryGraphics.java
@@ -15,7 +15,7 @@
  */
 
 
-package com.twosigma.beaker.chart.categoryplot.plotitems;
+package com.twosigma.beaker.chart.categoryplot.plotitem;
 
 import com.twosigma.beaker.chart.Color;
 import com.twosigma.beaker.chart.Graphics;
