+++ plugin/jvm/src/main/java/com/twosigma/beaker/chart/categoryplot/plotitem/CategoryPoints.java
@@ -15,8 +15,8 @@
  */
 
 
-package com.twosigma.beaker.chart.categoryplot.plotitems;
+package com.twosigma.beaker.chart.categoryplot.plotitem;
 
 
-public class CategoryLines extends CategoryGraphics {
+public class CategoryPoints extends CategoryGraphics {
 }
