+++ plugin/jvm/src/main/java/com/twosigma/beaker/chart/categoryplot/plotitem/CategoryStems.java
@@ -15,7 +15,7 @@
  */
 
 
-package com.twosigma.beaker.chart.categoryplot.plotitems;
+package com.twosigma.beaker.chart.categoryplot.plotitem;
 
 
 public class CategoryStems  extends CategoryGraphics {
