+++ plugin/jvm/src/main/java/com/twosigma/beaker/chart/serializer/CategoryBarsSerializer.java
@@ -16,10 +16,9 @@
 
 package com.twosigma.beaker.chart.serializer;
 
-import com.twosigma.beaker.chart.categoryplot.plotitems.CategoryBars;
+import com.twosigma.beaker.chart.categoryplot.plotitem.CategoryBars;
 import org.codehaus.jackson.JsonGenerator;
 import org.codehaus.jackson.JsonProcessingException;
-import org.codehaus.jackson.map.JsonSerializer;
 import org.codehaus.jackson.map.SerializerProvider;
 
 import java.io.IOException;
