+++ plugin/jvm/src/main/java/com/twosigma/beaker/chart/serializer/CategoryLinesSerializer.java
@@ -16,7 +16,7 @@
 
 package com.twosigma.beaker.chart.serializer;
 
-import com.twosigma.beaker.chart.categoryplot.plotitems.CategoryLines;
+import com.twosigma.beaker.chart.categoryplot.plotitem.CategoryLines;
 import org.codehaus.jackson.JsonGenerator;
 import org.codehaus.jackson.JsonProcessingException;
 import org.codehaus.jackson.map.SerializerProvider;
