+++ plugin/jvm/src/main/java/com/twosigma/beaker/chart/serializer/CategoryStemsSerializer.java
@@ -16,7 +16,7 @@
 
 package com.twosigma.beaker.chart.serializer;
 
-import com.twosigma.beaker.chart.categoryplot.plotitems.CategoryStems;
+import com.twosigma.beaker.chart.categoryplot.plotitem.CategoryStems;
 import org.codehaus.jackson.JsonGenerator;
 import org.codehaus.jackson.JsonProcessingException;
 import org.codehaus.jackson.map.SerializerProvider;
