+++ plugin/jvm/src/main/java/com/twosigma/beaker/jvm/module/SerializerModule.java
@@ -24,10 +24,10 @@ import com.twosigma.beaker.BeakerProgressUpdate;
 import com.twosigma.beaker.NamespaceClient;
 import com.twosigma.beaker.chart.Color;
 import com.twosigma.beaker.chart.categoryplot.CategoryPlot;
-import com.twosigma.beaker.chart.categoryplot.plotitems.CategoryBars;
-import com.twosigma.beaker.chart.categoryplot.plotitems.CategoryLines;
-import com.twosigma.beaker.chart.categoryplot.plotitems.CategoryPoints;
-import com.twosigma.beaker.chart.categoryplot.plotitems.CategoryStems;
+import com.twosigma.beaker.chart.categoryplot.plotitem.CategoryBars;
+import com.twosigma.beaker.chart.categoryplot.plotitem.CategoryLines;
+import com.twosigma.beaker.chart.categoryplot.plotitem.CategoryPoints;
+import com.twosigma.beaker.chart.categoryplot.plotitem.CategoryStems;
 import com.twosigma.beaker.chart.legend.LegendPosition;
 import com.twosigma.beaker.chart.serializer.AreaSerializer;
 import com.twosigma.beaker.chart.serializer.BarsSerializer;
